+++ CHANGELOG.md
@@ -6,7 +6,7 @@ More infos about this file : http://keepachangelog.com/
 
 ## [Unreleased] - no_due_date
 
-- **Update approvals-java to 0.11.1**
+- **Change approvals-java 0.7.1 to approvals-scala 0.12.1**
 
 ## [v1.1.8] - 2019-11-25